Rewrite all code to its definitive version
This change adds the definitive version of the robot code.
This version is made to be as flexible as possible for the design of the
questions.
It also supports branching as it's data structure depends on a set of
addressable HashMaps:

 > Pools, sets of questions that will be chosen at random when a pool is
	    addressed. Their prefix is: ?
 > Loose questions, no prefix.
 > Video transitions/looping videos. Their prefix is: !

The HashMaps are populated using a .json file, so editting code is not
necesarry for adding simple questions.

Each question has the name of it's target for each answer,
Each pool has the name of a target for when the pool is empty or the max
questions have been popped.
Each video has a target for when it ends or is forcibly stopped.

The .json file also enables changing of variables on a particular answer
of a particular question. This can be used to post custom end messages
and more.

I hope this version will be more or less sufficient to make a good
questionaire.